pyproject.toml
[project]
name = "temporal-lanchain"
version = "0.1.0"
description = "Temporal + LangChain integration sample with unified Langfuse tracing via OpenTelemetry"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-openai>=0.3.28",
    "langfuse>=3.2.1",
    "python-dotenv>=1.1.1",
    "temporalio>=1.14.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pylint>=3.3.7",
]
